requirements.txt
streamlit>=1.32.0
anthropic>=0.34.0
google-genai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
python-dotenv>=1.0.0
lxml>=5.0.0
markdown>=3.5.0
